refactor: move breakpoint types to a mod
