+++ src/ptrace.rs
@@ -1,3 +1,4 @@
+mod breakpoint;
 mod engine;
 mod guards;
 mod inspect;
@@ -5,10 +6,9 @@ mod syscall;
 mod tracer;
 mod waitpid;
 
+pub use breakpoint::*;
 pub use engine::RecursivePtraceEngine;
 pub use guards::*;
 pub use inspect::InspectError;
-pub use tracer::{
-  BreakPoint, BreakPointHit, BreakPointPattern, BreakPointStop, BreakPointType, Tracer,
-};
+pub use tracer::Tracer;
 pub use waitpid::*;
